Extracted filesystem utils to a separate file.
